+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ### Bug fixes
 * Shader optimizer skips parsed shaders that are null or parsed incorrectly.
 * Fix floats getting written with system localization instead of invariant culture.
+* Fix crash when an animation has a material swap that doesn't point to an transform on the avatar. Fix for: [Null pointer error on Create Optimized Copy](https://github.com/d4rkc0d3r/d4rkAvatarOptimizer/issues/7)
 
 ### Debug Info
 * Catch exception in the shader optimizer to show which shader is causing the exception.